Change invalid typespec error to warning
Raising a compiler error was correct because invalid typespecs were
generated. Unfortunately it also broke code that compiled and ran
without issues because the issue would only be noticed when running
dialyzer. To avoid breaking working code the error is changed to a
warning instead.
